Voice_RAG_with_GeminiAI_SDK/requirements.txt
# Core app
streamlit
python-dotenv

# Vector DB
qdrant-client
fastembed

# LangChain utilities
langchain
langchain-community

# Gemini (Google Generative AI SDK)
google-generativeai
